# api_service/app/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.DataResponse:
    properties:
      count:
        type: integer
      data: {}
    type: object
...
  models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.IdResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  models.PutgDTO:
    properties:
      coating:
        items:
          type: string
        type: array
      construction:
        items:
          $ref: '#/definitions/proto.PutgConstructions'
        type: array
      flangeId:
        type: string
      form:
        type: string
      graphite:
        items:
          type: string
        type: array
      iLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      mounting:
        items:
          type: string
        type: array
      oLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      obturator:
        $ref: '#/definitions/proto.PutgMaterials'
      reinforce:
        $ref: '#/definitions/proto.PutgMaterials'
      temperatures:
        items:
          $ref: '#/definitions/proto.PutgTemp'
        type: array
      typeFlId:
        type: string
      typePr:
        type: string
    required:
    - coating
    - construction
    - flangeId
    - form
    - graphite
    - mounting
    - temperatures
    - typeFlId
    - typePr
    type: object
...
  models.PutgmDTO:
    properties:
      basis:
        $ref: '#/definitions/proto.PutgMaterials'
      coating:
        items:
          type: string
        type: array
      construction:
        items:
          $ref: '#/definitions/proto.PutgmConstructions'
        type: array
      flangeId:
        type: string
      form:
        type: string
      graphite:
        items:
          type: string
        type: array
      mounting:
        items:
          type: string
        type: array
      obturator:
        $ref: '#/definitions/proto.PutgMaterials'
      temperatures:
        items:
          $ref: '#/definitions/proto.PutgTemp'
        type: array
      typeFlId:
        type: string
      typePr:
        type: string
    required:
    - coating
    - construction
    - flangeId
    - form
    - graphite
    - mounting
    - temperatures
    - typeFlId
    - typePr
    type: object
...
  proto.ConTemp:
    properties:
      constructions:
        items:
          $ref: '#/definitions/proto.Constr'
        type: array
      temp:
        type: string
    type: object
  proto.Constr:
    properties:
      obturators:
        items:
          $ref: '#/definitions/proto.PutgObt'
        type: array
      short:
        type: string
    type: object
...
  proto.Putg:
    properties:
      coating:
        items:
          type: string
        type: array
      construction:
        items:
          $ref: '#/definitions/proto.PutgConstructions'
        type: array
      form:
        type: string
      graphite:
        items:
          type: string
        type: array
      iLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      id:
        type: string
      mounting:
        items:
          type: string
        type: array
      oLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      obturator:
        $ref: '#/definitions/proto.PutgMaterials'
      reinforce:
        $ref: '#/definitions/proto.PutgMaterials'
      temperatures:
        items:
          $ref: '#/definitions/proto.PutgTemp'
        type: array
      typeFlId:
        type: string
      typePr:
        type: string
    type: object
  proto.PutgConstructions:
    properties:
      grap:
        type: string
      temperatures:
        items:
          $ref: '#/definitions/proto.ConTemp'
        type: array
    type: object
...
  proto.PutgMaterials:
    properties:
      default:
        type: string
      obturators:
        items:
          type: string
        type: array
      values:
        items:
          type: string
        type: array
    type: object
  proto.PutgObt:
    properties:
      imageUrl:
        type: string
      short:
        type: string
    type: object
  proto.PutgTemp:
    properties:
      grap:
        type: string
      temps:
        items:
          $ref: '#/definitions/proto.Temp'
        type: array
    type: object
  proto.Putgm:
    properties:
      basis:
        $ref: '#/definitions/proto.PutgMaterials'
      coating:
        items:
          type: string
        type: array
      construction:
        items:
          $ref: '#/definitions/proto.PutgmConstructions'
        type: array
      form:
        type: string
      graphite:
        items:
          type: string
        type: array
      id:
        type: string
      mounting:
        items:
          type: string
        type: array
      obturator:
        $ref: '#/definitions/proto.PutgMaterials'
      temperatures:
        items:
          $ref: '#/definitions/proto.PutgTemp'
        type: array
      typeFlId:
        type: string
      typePr:
        type: string
    type: object
  proto.PutgmConstr:
    properties:
      basis:
        type: string
      obturator:
        items:
          $ref: '#/definitions/proto.PutgmObt'
        type: array
    type: object
  proto.PutgmConstructions:
    properties:
      basis:
        items:
          $ref: '#/definitions/proto.PutgmConstr'
        type: array
      grap:
        type: string
    type: object
...
  proto.PutgmObt:
    properties:
      obturator:
        type: string
      sealant:
        items:
          $ref: '#/definitions/proto.PutgmSeal'
        type: array
    type: object
  proto.PutgmSeal:
    properties:
      imageUrl:
        type: string
      seal:
        type: string
    type: object
...
  proto.Temp:
    properties:
      id:
        type: string
      mods:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Service for Sealur
  title: Sealur
  version: 0.14.0
paths:
...
  /sealur-pro/putg:
    get:
      consumes:
      - application/json
      description: получение прокладок путг
      parameters:
      - description: form
        in: query
        name: form
        required: true
        type: string
      - description: flange id
        in: query
        name: flangeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.Putg'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Putg
      tags:
      - Sealur Pro -> putg
    post:
      consumes:
      - application/json
      description: создание прокладки путг
      parameters:
      - description: putg info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Putg
      tags:
      - Sealur Pro -> putg
...
  /sealur-pro/putg/{id}:
    delete:
      consumes:
      - application/json
      description: удаление прокладки путг
      parameters:
      - description: putg id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Putg
      tags:
      - Sealur Pro -> putg
    put:
      consumes:
      - application/json
      description: обновление прокладки путг
      parameters:
      - description: putg info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgDTO'
      - description: putg id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Putg
      tags:
      - Sealur Pro -> putg
  /sealur-pro/putgm:
    get:
      consumes:
      - application/json
      description: получение прокладок путгм
      parameters:
      - description: form
        in: query
        name: form
        required: true
        type: string
      - description: flange id
        in: query
        name: flangeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.Putgm'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Putgm
      tags:
      - Sealur Pro -> putgm
    post:
      consumes:
      - application/json
      description: создание прокладки путгм
      parameters:
      - description: putgm info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgmDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Putgm
      tags:
      - Sealur Pro -> putgm
...
  /sealur-pro/putgm/{id}:
    delete:
      consumes:
      - application/json
      description: удаление прокладки путгм
      parameters:
      - description: putgm id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Putgm
      tags:
      - Sealur Pro -> putgm
    put:
      consumes:
      - application/json
      description: обновление прокладки путгм
      parameters:
      - description: putgm info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgmDTO'
      - description: putgm id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Putgm
      tags:
      - Sealur Pro -> putgm
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"